WebServer/flaskr/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
APScheduler==3.10.0
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==3.6.4.0
blinker==1.5
Brotli==1.0.9
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
google-cloud-storage==1.44.0
importlib-metadata==6.0.0
inflate64==0.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
multivolumefile==0.2.3
packaging==23.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary
py7zr==0.20.4
pybcj==1.0.1
pycryptodomex==3.17
PyJWT==2.6.0
pyppmd==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyzstd==0.15.4
six==1.16.0
SQLAlchemy==2.0.1
texttable==1.6.7
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.14.0